AtgDsp: RmsCore window template param long (match DelayLine<T,long N>)
clang rejects deducing the DWORD (unsigned long) non-type template parameter
from a DelayLine instantiated with a long extent -- 'long vs unsigned long',
which MSVC tolerates. RmsCore's window size is DelayLine<__vector4, windowLen>,
so windowLen must be long like DelayLine's N. Fixes XAudio2CustomAPO.

+++ Common/AtgDsp.h
@@ -133,7 +133,7 @@ private:
 // be initialized to zero.
 //
 //--------------------------------------------------------------------------------------
-template<DWORD windowLen> __vector4 __forceinline RmsCore( vector4_in input, DelayLine<__vector4, windowLen>* __restrict window, vector4_inout accumulator, __vector4 oneOverWindowLen )
+template<long windowLen> __vector4 __forceinline RmsCore( vector4_in input, DelayLine<__vector4, windowLen>* __restrict window, vector4_inout accumulator, __vector4 oneOverWindowLen )
  {
     // 
     // Square the four newest input samples
